+++ examples/react-query/src/gen/hooks/pet/useDeletePetHook.ts
@@ -1,7 +1,7 @@
 import client from '@kubb/plugin-client/clients/axios'
 import type { DeletePetMutationResponse, DeletePetPathParams, DeletePetHeaderParams, DeletePet400 } from '../../models/DeletePet.ts'
 import type { RequestConfig, ResponseErrorConfig } from '@kubb/plugin-client/clients/axios'
-import type { UseMutationOptions } from '@tanstack/react-query'
+import type { UseMutationOptions, QueryClient } from '@tanstack/react-query'
 import { useMutation } from '@tanstack/react-query'
 
 export const deletePetMutationKey = () => [{ url: '/pet/{pet_id}' }] as const
@@ -41,11 +41,14 @@ export function useDeletePetHook<TContext>(
       ResponseErrorConfig<DeletePet400>,
       { pet_id: DeletePetPathParams['pet_id']; headers?: DeletePetHeaderParams },
       TContext
-    >
+    > & { client?: QueryClient }
     client?: Partial<RequestConfig> & { client?: typeof client }
   } = {},
 ) {
-  const { mutation: mutationOptions, client: config = {} } = options ?? {}
+  const {
+    mutation: { client: queryClient, ...mutationOptions } = {},
+    client: config = {},
+  } = options ?? {}
   const mutationKey = mutationOptions?.mutationKey ?? deletePetMutationKey()
 
   return useMutation<
@@ -53,11 +56,14 @@ export function useDeletePetHook<TContext>(
     ResponseErrorConfig<DeletePet400>,
     { pet_id: DeletePetPathParams['pet_id']; headers?: DeletePetHeaderParams },
     TContext
-  >({
-    mutationFn: async ({ pet_id, headers }) => {
-      return deletePetHook({ pet_id }, headers, config)
+  >(
+    {
+      mutationFn: async ({ pet_id, headers }) => {
+        return deletePetHook({ pet_id }, headers, config)
+      },
+      mutationKey,
+      ...mutationOptions,
     },
-    mutationKey,
-    ...mutationOptions,
-  })
+    queryClient,
+  )
 }